working with localhost

+++ src/main/kotlin/haileyj/wordofmouthbackend/entity/Company.kt
@@ -21,7 +21,7 @@ data class Company(
     var id: String = ""
 
     @Transient
-    var rating: Double? = null
+    var rating: Float? = null
 
     @Transient
     var reviews: List<Review> = ArrayList()
